Currency need to accept Quantity as double and not int
Currency need to accept Quantity as double and not int

+++ app/src/main/java/br/com/carteira/data/PortfolioProvider.java
@@ -1287,7 +1287,7 @@ public class PortfolioProvider extends ContentProvider {
 
         final SQLiteDatabase db = dbHelper.getWritableDatabase();
 
-        int quantity;
+        double quantity;
         double currentPrice;
         double totalBuy;
         double incomeTotal;
@@ -1312,7 +1312,7 @@ public class PortfolioProvider extends ContentProvider {
                     queryCursor.moveToFirst();
 
                     currentPrice = Double.parseDouble(contValues.get(key).toString());
-                    quantity = queryCursor.getInt(queryCursor.getColumnIndex(PortfolioContract
+                    quantity = queryCursor.getDouble(queryCursor.getColumnIndex(PortfolioContract
                             .CurrencyData.COLUMN_QUANTITY_TOTAL));
                     totalBuy = queryCursor.getDouble(queryCursor.getColumnIndex(PortfolioContract
                             .CurrencyData.COLUMN_BUY_VALUE_TOTAL));
